updated byteCollection creation
