--- academy-ui/src/types/response/ByteCollectionWithApiResponses.ts
@@ -1,8 +1,0 @@
-export interface CreateByteCollectionWithApiResponse {
-  byteCollection: {
-    id: string;
-    name: string;
-    description: string;
-    byteIds: Array<string>;
-  }
-}
